Render slides from URL parameter source path in Hugo
To render markdown js in an iframe in a hugo page, the index.html is modified
to read the source path from the url parameter passed to the link in the iframe.

This allows us to specify the path to the source in the content page of hugo
and allows us to embed slideshows easily and keep its content separated.